[DF-8140] details from review fixed
Former-commit-id: 1a70e0de2fee4fe26be6d8aa95d57a763e25e2f1

+++ Tools/QuickEd/Classes/Document.cpp
@@ -146,7 +146,7 @@ void Document::SetPixelization(bool hasPixelization)
 
 void Document::SetDPR(qreal arg)
 {
-    double dpr = static_cast<double>(arg);
+    float32 dpr = static_cast<float32>(arg);
     systemManager.DPRChanged.Emit(dpr);
 }
